requirements.txt
discord.py
imgurpython
peewee==3.6.4